refactor: MenuGroup 도메인 Entity로 변경
