+++ src/main/java/kitchenpos/domain/menugroup/MenuGroup.java
@@ -1,0 +1,30 @@
+package kitchenpos.domain.menugroup;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Entity
+public class MenuGroup {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String name;
+
+    protected MenuGroup() {
+    }
+
+    public MenuGroup(String name) {
+        this.name = name;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
